1/6/25 @ 6:58 PM -
Changed "add" command so that an "active" flag defaulted at false is sent to the generated profile's json file upon execution.  This flag will be used by the load command to restrict the maximum number of active profiles to 10; promoting resource efficiency.

Will update the "load" command to generate a directory within resources\assets\profile{index#} called ""{profile name}\frames".  This is where the frames ("screenshots") generated by the "web screen capture client" will be sent.

The profile name is passed as an argument for the load command, and the mod will create a directory unique to that profile.  It will also create its own unique call for the web screen capture client.  10 active profiles can exist at once, but this is subject to revision.