1/7/25 @ 4:30 PM
Screenshot.js is now capable of being called from inside of Minecraft.  A process of it is created, passing the profile name as an argument.

Current issue - duplicate functionality in load.java and in screenshot.js.  Screenshot.js is being edited so that it accepts both <frame target directory> and <name> as a directory.  Load will pass the correct arguments and create a process of screenshot.js accordingly; newly passing the target directory as an argument for the process creation of screenshot.js.

The directory was also significantly reorganized.  node's files are no longer located in the root directory where they'd be compressed by the built jar file and inaccessible.  Later, some editing to build.gradle will be necessary to declare how the mod will acquire node, puppeteer, and any other necessary dependencies.