1/7/25 @ 4:57
Big milestone! The web renderer is now executable from inside of Minecraft!

Logic will need to be implemented in remove.java in order to end the related renderer processes when removing a profile.  As of right now, the renderer task continues indefinitely.  Ending it will preserve unnecessary resource usage.

This implies the possibility of an unload command as well, if I end up deciding it's worthwhile.  This unload functionality could later be accessed from other areas to end the renderer process if no players are within the simulation distance of the created map wall.