1/7/25 @ 9:13 PM
Unload command has been successfully initialized in game and is fully functional.  Screenshot.js was edited to clean itself up when the profile it's runnin for's json file has its active status updated to false, which is done by the unload command.

My plans now are to implement the unload functionality into the remove class.  This way, if a player executes the remove command, the render buffers clean themselves up before the profile is removed.  This mitigates the risk of the user not being able to add a mapcast if the buffers aren't removed, since several functionalities depend on an empty framecontainer directory.

Just need to fix remove to implement this.  Once that is successful, the main branch will be updated and I'll start integrating the unload functionality to world exits and absence of players within simulation distance (or the max simulation distance configured by the server)

Some changes have also been made to the generation of json files, allowing originx, y, z, and facing directions as arguments for the load command, which will allow the mod to know where mapcasts are loaded so it can unload them outside of simulation distance.